don't restore folds if they are already present
This problem manifests itself when files are reverted. Overlays are not
deleted in this case and ‘vimish-fold’ tries to create overlays on tops
of already existing ones — this results in error message “Fold already
exists here”. Since fold configuration before reverting may be way
different from saved one, we should prefer it and not attempt
restoration from file in these cases.